Update tooltip logic
