+++ src/components/ZkProofButton.tsx
@@ -1,7 +1,4 @@
 import { AccentText } from 'components/Text'
-import { MutableRef } from 'preact/hooks'
-import { useRef } from 'react'
-import { useState } from 'preact/hooks'
 import ToolTip from 'components/ToolTip'
 import classnames, {
   alignItems,
@@ -17,27 +14,15 @@ import classnames, {
   justifyContent,
   margin,
   padding,
-  position,
   space,
   width,
 } from 'classnames/tailwind'
-import useClickOutside from 'hooks/useClickOutside'
-
-const zkProofButtonTooltip = classnames(
-  position('relative'),
-  width('w-mobile-card', 'sm:w-card')
-)
-const tooltipWrapper = classnames(
-  position('relative'),
-  height('h-px'),
-  width('w-full'),
-  margin('mx-auto')
-)
 
 const zkProofButton = classnames(
   display('flex'),
   flexDirection('flex-row'),
   justifyContent('justify-center'),
+  width('w-mobile-card', 'sm:w-card'),
   backgroundColor('bg-primary-background'),
   space('space-x-2'),
   padding('py-4'),
@@ -63,31 +48,17 @@ const QuestionMark = () => {
 }
 
 export default function () {
-  const [isShow, setIsShow] = useState(false)
-  const zkProofButtonRef = useRef() as MutableRef<HTMLDivElement>
-
   const popoverText =
     'In your wallet(s), you have NFTs that can point back to your identity (aka, getting doxxed). But with ZK proof, you can verify ownership of NFTs while staying pseudonymous.'
 
-  useClickOutside(zkProofButtonRef, () => setIsShow(false))
-
   return (
-    <div className={zkProofButtonTooltip}>
-      <div className={tooltipWrapper}>
-        <ToolTip text={popoverText} show={isShow} />
-      </div>
-      <div
-        ref={zkProofButtonRef}
-        className={zkProofButton}
-        onMouseEnter={() => setIsShow(true)}
-        onMouseLeave={() => setIsShow(false)}
-        onClick={() => setIsShow(true)}
-      >
+    <ToolTip text={popoverText}>
+      <div className={zkProofButton}>
         <AccentText color="text-accent">What's a ZK proof?</AccentText>
         <AccentText small primary color="text-accent">
           <QuestionMark />
         </AccentText>
       </div>
-    </div>
+    </ToolTip>
   )
 }
